requirements.txt
openai
python-dotenv
requests
schedule
apscheduler
